requirements.txt
# FlatSat Device Simulator Requirements
# Comprehensive multi-device simulator for ARS, Magnetometer, and Reaction Wheel

# Core dependencies
python-can>=4.0.0          # CAN bus communication
pyserial>=3.5               # RS485/Serial communication
# crcmod>=1.7               # CRC calculation - replaced with built-in implementation
flask>=2.0.0                # REST API web framework
requests>=2.25.0            # HTTP client for API testing

# Optional dependencies for enhanced functionality
numpy>=1.21.0               # Numerical operations for calibration
matplotlib>=3.5.0           # Data visualization
pandas>=1.3.0               # Data analysis and logging

# Development and testing dependencies
pytest>=6.0.0               # Testing framework
pytest-cov>=2.12.0          # Coverage reporting
black>=21.0.0                # Code formatting
flake8>=3.9.0                # Linting

# Documentation dependencies
sphinx>=4.0.0                # Documentation generation
sphinx-rtd-theme>=0.5.0      # Read the Docs theme

# Logging and monitoring
colorlog>=6.6.0              # Colored logging output
